apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
httpx>=0.27.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.40.0
replicate>=0.25.0
pyspellchecker>=0.8.0
supabase>=2.0.0
